requirements.txt
requests
python-consul
six
